requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.44
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2  # Direct bcrypt library (more reliable than passlib for Python 3.13)
python-dotenv==1.0.0

# AI Integration (using HuggingFace as primary - FREE, OpenAI as optional fallback)
# Note: HuggingFace API doesn't require torch locally, it runs on their servers
# OpenAI is optional - only needed if you want to use it as fallback
# openai==1.3.5  # Uncomment if you want OpenAI support

# HTTP client for async requests
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0

